add bash alias

+++ install.sh
@@ -6,13 +6,26 @@ echo "cd to .vim directory"
 cd ~/.vim
 
 echo "create .vimrc"
+if [ -f ~/.vimrc ]; then
 rm ~/.vimrc
-rm ~/.gvimrc
+fi
 ln -s ~/.vim/others/vimrc.local ~/.vimrc
+if [ -f ~/.grvimrc ]; then
+rm ~/.gvimrc
+fi
 ln -s ~/.vim/others/vimrc.local ~/.gvimrc
 
-rm ~/.screenrc
+if [ -f ~/.screenrc ]; then
+    rm ~/.screenrc
+fi
 ln -s ~/.vim/others/screenrc ~/.screenrc
+
+if [ -f ~/.bash_aliases ]; then
+    rm ~/.bash_aliases
+fi
+ln -s ~/.vim/others/bash_aliases ~/.bash_aliases
+
+echo ". ~/.bash_aliases" >> ~/.bashrc
 
 mkdir ~/bin
 rm ~/bin/mkc.sh
